Updated for release
The RockBlock was unable to send the message with the lid closed. I changed the RockBlock config to use the USB profile since the power is coming through the Ardunio. 
Also commented out getting time from RockBlock. The hour of the day being returned was off from the expected UTC time.